+++ src/test/java/com/openggf/trace/TraceFixtures.java
@@ -71,6 +71,7 @@ public final class TraceFixtures {
                 null,
                 null,
                 null,
+                null,
                 null);
     }
 }